Moving the parser files ot a better location
