+++ src/I18n/Parser/MoFileParser.php
